+++ src/extension.ts
@@ -4,36 +4,55 @@ import * as vscode from 'vscode';
 import { GycTools } from './GycTools';
 import { CodeGenerator } from './CodeGenerator';
 import { FileExplorer } from './fileExplorer';
+import { DocumentGenerator } from './DocumentGenerator';
+import { DataBaseCodeGenerator } from './DataBaseCodeGenerator';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
 
 	try {
-		GycTools.Utils.inistializeTemplate(context);
-
-		let targetProject: any = GycTools.Utils.getTargetProjectConfig(context);
-
+		const tpConig = vscode.workspace.getConfiguration().get<string>('gyctools.templatePath','');
+		let targetProject = GycTools.Utils.getTargetProjectConfig(context);
+		if ( !targetProject || !targetProject.templatePath || targetProject.templatePath.length <= 0) {
+			GycTools.Utils.inistializeTemplate(context,tpConig);
+		}
+		if (!targetProject) {
+			GycTools.Utils.initializeConfigFile(context);
+			targetProject = GycTools.Utils.getTargetProjectConfig(context);
+		} else { //get new config alltimes
+			targetProject = GycTools.Utils.getTargetProjectConfig(context);
+		}
+		
 		let templateExplorer: FileExplorer = new FileExplorer(context, GycTools.Utils.getTemplatePath(context, targetProject));
 
-		let initPrject = function () {
-			if (!targetProject) {
-				GycTools.Utils.initializeConfigFile(context);
-				targetProject = GycTools.Utils.getTargetProjectConfig(context);
-				templateExplorer.reFlashRoot();
-				return;
-			} else { //get new config alltimes
-				targetProject = GycTools.Utils.getTargetProjectConfig(context);
-			}
-			if (targetProject === undefined) {
-				console.error('gyctool config error , extension disabled');
-				vscode.window.showErrorMessage('gyctool config error , extension disabled');
-				return;
-			}
-		};
+		templateExplorer.reFlashRoot();
+		if (targetProject === undefined) {
+			console.error('gyctool config error , extension disabled');
+			return;
+		}
+
+		// let initProject = function () {
+		// 	targetProject = GycTools.Utils.getTargetProjectConfig(context);
+		// };
+
+		context.subscriptions.push(vscode.commands.registerCommand('gyctools.generateDatabaseSpecification', (item) => {
+			targetProject = GycTools.Utils.getTargetProjectConfig(context);
+			const docGenerator: DocumentGenerator = new DocumentGenerator(context, item, targetProject);
+			docGenerator.generateNew();
+			return;
+		}));
+
+        context.subscriptions.push(vscode.commands.registerCommand('gyctools.generateCodeByAllTable', (item) => {
+			targetProject = GycTools.Utils.getTargetProjectConfig(context);
+			const dbGenerator: DataBaseCodeGenerator = new DataBaseCodeGenerator(context, item, targetProject);
+			const count = dbGenerator.generateNew();
+            vscode.window.showInformationMessage('Generate ' + count + ' Code Files', 'OK');
+			return;
+		}));
 
 		context.subscriptions.push(vscode.commands.registerCommand('gyctools.generateYourCodes', (item) => {
-			initPrject();
+			targetProject = GycTools.Utils.getTargetProjectConfig(context);
 			const codeGenerator: CodeGenerator = new CodeGenerator(context, item, targetProject);
 			codeGenerator.generateNew();
 			return;
@@ -41,13 +60,7 @@ export function activate(context: vscode.ExtensionContext) {
 	
 	} catch (error) {
 		console.error(error);
-		vscode.window.showErrorMessage('gyc activate failed',error);
 	}
-	
-
-
-
-
 	
 
 }